Rework logging
- Move logging code to dedicated source files
- Disable colour output when the output is not a terminal
- Add support for logging to a file
- Make it possible to select the output (stdout, stderr or file for now)
- Remove a few globals
- Fix a global init race condition (init_logging vs. gLoggingLevel)

Fixes #99
